+++ src/themes.tsx
@@ -3,10 +3,10 @@ import { createRoot } from 'react-dom/client'
 import './index.css'
 import './styles/themes.css'
 import './styles/pivot.css'
-import { DemoApp } from './demo/DemoApp'
+import { ThemesPage } from './demo/ThemesPage'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <DemoApp />
+    <ThemesPage />
   </StrictMode>,
 )